Move callback handling to entity class
